+++ R/exclude.R
@@ -22,16 +22,15 @@ exclude <- function(lints, exclusions = settings$exclusions, ...) {
   source_exclusions <- lapply(filenames, parse_exclusions, ...)
   names(source_exclusions) <- filenames
 
-  excl <- normalize_exclusions(c(source_exclusions, exclusions))
-  excl_names <- names(excl)
 
+  exclusions <- normalize_exclusions(c(source_exclusions, exclusions))
   to_exclude <- vapply(seq_len(nrow(df)),
     function(i) {
       file <- df$filename[i]
-      file %in% excl_names &&
-        excl[[file]] == Inf ||
-        df$line_number[i] %in% excl[[file]]
-    },
+      file %in% names(exclusions) &&
+        exclusions[[file]] == Inf ||
+        df$line_number[i] %in% exclusions[[file]]
+     },
     logical(1))
 
   if (any(to_exclude)) {
@@ -69,7 +68,7 @@ parse_exclusions <- function(file, exclude = settings$exclude,
   sort(unique(c(exclusions, which(rex::re_matches(lines, exclude)))))
 }
 
-normalize_exclusions <- function(x) {
+normalize_exclusions <- function(x, normalize_path=TRUE) {
   if (is.null(x) || length(x) <= 0) {
     return(list())
   }
@@ -97,6 +96,11 @@ normalize_exclusions <- function(x) {
       names(x)[unnamed] <- x[unnamed]
       x[unnamed] <- Inf
     }
+  }
+
+  if (normalize_path) {
+    x <- x[file.exists(names(x))]       # remove exclusions for non-existing files
+    names(x) <- normalizePath(names(x)) # get full path for remaining files
   }
 
   remove_line_duplicates(